Move entity directory to entities
Make the directory plural to map to the other directory names that
store differnt file types.
