+++ ormconfig.json
@@ -8,7 +8,7 @@
    "synchronize": false,
    "logging": false,
    "entities": [
-      "dist/src/entity/**/*.js"
+      "dist/src/entities/**/*.js"
    ],
    "migrations": [
       "dist/src/migrations/**/*.js"
@@ -17,7 +17,7 @@
       "dist/src/subscriber/**/*.js"
    ],
    "cli": {
-      "entitiesDir": "src/entity",
+      "entitiesDir": "src/entities",
       "migrationsDir": "src/migrations",
       "subscribersDir": "src/subscriber"
    }
